Rename getInstance to createExecutor in CompactionExecutorFactory. Moved Rebalance compaction checks from Worker to AcidCompactionExecutor. Moved CompactionTxn from Worker to AcidCompactionExecutor because it is ACID specific. Removed collectMrStats from IcebergCompactionExecutor. Added a test to verify that subsequent compaction on Iceberg table works as expected.
