+++ iceberg/iceberg-handler/src/test/queries/positive/iceberg_major_compaction_partitioned.q
@@ -56,3 +56,36 @@ alter table ice_orc COMPACT 'major' and wait;
 select * from ice_orc;
 describe formatted ice_orc;
 show compactions;
+
+-- Starting second set of inserts/updates/deletes and calling compaction at the end
+-- to check that subsequent compaction works
+
+insert into ice_orc VALUES ('fn11','ln11', 1);
+insert into ice_orc VALUES ('fn12','ln12', 1);
+insert into ice_orc VALUES ('fn13','ln13', 1);
+insert into ice_orc VALUES ('fn14','ln14', 1);
+insert into ice_orc VALUES ('fn15','ln15', 2);
+insert into ice_orc VALUES ('fn16','ln16', 2);
+insert into ice_orc VALUES ('fn17','ln17', 2);
+insert into ice_orc VALUES ('fn18','ln18', 2);
+
+update ice_orc set last_name = 'ln11a' where first_name='fn11';
+update ice_orc set last_name = 'ln12a' where first_name='fn12';
+update ice_orc set last_name = 'ln13a' where first_name='fn13';
+update ice_orc set last_name = 'ln14a' where first_name='fn14';
+update ice_orc set last_name = 'ln15a' where first_name='fn15';
+update ice_orc set last_name = 'ln16a' where first_name='fn16';
+update ice_orc set last_name = 'ln17a' where first_name='fn17';
+update ice_orc set last_name = 'ln18a' where first_name='fn18';
+
+delete from ice_orc where last_name in ('ln11a', 'ln12a', 'ln17a', 'ln18a');
+
+select * from ice_orc;
+describe formatted ice_orc;
+
+explain alter table ice_orc COMPACT 'major' and wait;
+alter table ice_orc COMPACT 'major' and wait;
+
+select * from ice_orc;
+describe formatted ice_orc;
+show compactions;
